Added image upload/download (storing at the disk). eshop-admin module…

+++ eshop-db/src/main/java/com/mvnikitin/eshop/repositories/BrandRepository.java
@@ -1,9 +1,9 @@
 package com.mvnikitin.eshop.repositories;
 
-import com.mvnikitin.eshop.entities.Brand;
+import com.mvnikitin.eshop.model.Brand;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BrandRepository extends JpaRepository<Brand, Long> {
+public interface BrandRepository extends JpaRepository<Brand, Integer> {
 }
