+++ eshop-db/src/main/java/com/mvnikitin/eshop/repositories/ProductRepository.java
@@ -1,9 +1,9 @@
 package com.mvnikitin.eshop.repositories;
 
-import com.mvnikitin.eshop.entities.Product;
+import com.mvnikitin.eshop.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository  extends JpaRepository<Product, Long> {
+public interface ProductRepository  extends JpaRepository<Product, Integer> {
 }
