+++ eshop-db/src/main/java/com/mvnikitin/eshop/repositories/RoleRepository.java
@@ -1,9 +1,9 @@
 package com.mvnikitin.eshop.repositories;
 
-import com.mvnikitin.eshop.entities.Role;
+import com.mvnikitin.eshop.model.Role;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepository extends CrudRepository<Role, Long> {
+public interface RoleRepository extends CrudRepository<Role, Integer> {
 }
